Port to 1.16 (+ fix pom build)
