+++ .gitignore
@@ -1,3 +1,5 @@
-+.gitignore
- +.idea/
- +/target/ 
+.gitignore
+.idea/
+/target
+/BuildTools
+*.iml
