+++ src/main/java/me/F_o_F_1092/AlmostFlatLandsReloaded/PluginManager/Math.java
