+++ src/main/java/me/F_o_F_1092/AlmostFlatLandsReloaded/PluginManager/VersionManager.java
@@ -16,11 +16,12 @@ public class VersionManager {
 		v1_11_R1,
 		v1_12_R1,
 		v1_13_R1, v1_13_R2,
+		v1_16_R1,
 		UNKNOWN
 	}
 	
 	public enum BungeeVersion {
-		v1_7, v1_8, v1_9, v1_0, v1_10, v1_11, v1_12, v1_13, UNKNOWN
+		v1_7, v1_8, v1_9, v1_0, v1_10, v1_11, v1_12, v1_13, v1_16, UNKNOWN
 	}
 	
 	public enum ServerType {
@@ -59,6 +60,8 @@ public class VersionManager {
 				VersionManager.bungeeVersion = BungeeVersion.v1_12;
 			} else if (version.startsWith("1.13")) {
 				VersionManager.bungeeVersion = BungeeVersion.v1_13;
+			} else if (version.startsWith("1.16")) {
+				VersionManager.bungeeVersion = BungeeVersion.v1_16;
 			} else {
 				VersionManager.bungeeVersion = BungeeVersion.UNKNOWN;
 			}
