+++ src/main/java/me/F_o_F_1092/AlmostFlatLandsReloaded/WorldGenerator/Util.java
@@ -1,12 +1,12 @@
 package me.F_o_F_1092.AlmostFlatLandsReloaded.WorldGenerator;
 
 import org.bukkit.World;
-import org.bukkit.craftbukkit.v1_13_R2.CraftChunk;
+import org.bukkit.craftbukkit.v1_16_R1.CraftChunk;
 
-import net.minecraft.server.v1_13_R2.Block;
-import net.minecraft.server.v1_13_R2.BlockPosition;
-import net.minecraft.server.v1_13_R2.Chunk;
-import net.minecraft.server.v1_13_R2.IBlockData;
+import net.minecraft.server.v1_16_R1.Block;
+import net.minecraft.server.v1_16_R1.BlockPosition;
+import net.minecraft.server.v1_16_R1.Chunk;
+import net.minecraft.server.v1_16_R1.IBlockData;
 
 public class Util {
 	
@@ -15,7 +15,7 @@ public class Util {
 		IBlockData ibd = b.getBlockData();
 		Chunk c = ((CraftChunk)world.getBlockAt(x, y, z).getChunk()).getHandle();
 
-		c.a(bp, ibd, true);
+		c.setType(bp, ibd, true);
     }
 
 }
